Im honestly unsure where i was going with this wrapper. I may abandon it as it sits in this wierd middle ground of usefull and useless.
I think originally this was meant to allow more modern (4.4+) OpenGL semantics
to be used with older OpenGL versions by sort of reimplementing the new features
just using older API functions.

However, this is arguably a bad move as all it does is hurts performance on both
old and new OpenGL versions.

On the other hand, I will always have a softspot for bringing RAII and friends to the OpenGL party.
Because of this I think this will just become a "wrapper" around existing OpenGL
functions instead of the current mode of extending functionality.